@@ -16,28 +16,27 @@
         }
     ],
     "require": {
-        "php": "^5.5|^7.0",
+        "php": "^7.1",
         "friendsofsymfony/oauth2-php": "~1.1",
-        "symfony/framework-bundle": "~2.8|~3.0|^4.0",
-        "symfony/security-bundle": "~2.8|~3.0|^4.0",
-        "symfony/dependency-injection": "^2.8|~3.0|^4.0",
-        "paragonie/random_compat": "^1|^2"
+        "symfony/framework-bundle": "~3.0|^4.0",
+        "symfony/security-bundle": "~3.0|^4.0",
+        "symfony/dependency-injection": "~3.0|^4.0"
     },
     "require-dev": {
-        "symfony/class-loader": "~2.8|~3.0|^4.0",
-        "symfony/yaml": "~2.8|~3.0|^4.0",
-        "symfony/form": "~2.8|~3.0|^4.0",
-        "symfony/console": "~2.8|~3.0|^4.0",
-        "symfony/templating": "~2.8|~3.0|^4.0",
-        "willdurand/propel-typehintable-behavior": "^1.0.4",
-        "propel/propel1": "^1.6.5",
-        "phing/phing": "~2.4",
-        "doctrine/mongodb-odm": "~1.0",
         "doctrine/doctrine-bundle": "~1.0",
+        "doctrine/mongodb-odm": "~1.0",
         "doctrine/orm": "~2.2",
+        "phing/phing": "~2.4",
+        "php-mock/php-mock-phpunit": "^1.1",
         "phpunit/phpunit": "~4.8|~5.0",
-        "symfony/phpunit-bridge": "~2.8|~3.0|^4.0",
-        "php-mock/php-mock-phpunit": "^1.1"
+        "propel/propel1": "^1.6.5",
+        "symfony/class-loader": "~3.0|^4.0",
+        "symfony/console": "~3.0|^4.0",
+        "symfony/form": "~3.0|^4.0",
+        "symfony/phpunit-bridge": "~3.0|^4.0",
+        "symfony/templating": "~3.0|^4.0",
+        "symfony/yaml": "~3.0|^4.0",
+        "willdurand/propel-typehintable-behavior": "^1.0.4"
     },
     "suggest": {
         "doctrine/doctrine-bundle": "*",
@@ -51,9 +50,12 @@
         "psr-4": { "FOS\\OAuthServerBundle\\": "" },
         "exclude-from-classmap": ["/Tests/"]
     },
+    "config": {
+        "sort-packages": true
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.5-dev"
+            "dev-master": "2.0-dev"
         }
     }
 }
