@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the FOSOAuthServerBundle package.
+ *
+ * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace FOS\OAuthServerBundle\Tests\DependencyInjection\Compiler;
 
 use FOS\OAuthServerBundle\DependencyInjection\Compiler\RequestStackCompilerPass;
@@ -8,11 +19,11 @@ use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
- * Class RequestStackCompilerPassTest
- * @package FOS\OAuthServerBundle\Tests\DependencyInjection\Compiler
+ * Class RequestStackCompilerPassTest.
+ *
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class RequestStackCompilerPassTest extends \PHPUnit_Framework_TestCase
+class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var RequestStackCompilerPass
@@ -87,6 +98,4 @@ class RequestStackCompilerPassTest extends \PHPUnit_Framework_TestCase
 
         $this->assertNull($this->instance->process($this->container));
     }
-
-
 }
