@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the FOSOAuthServerBundle package.
+ *
+ * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace FOS\OAuthServerBundle\Tests\DependencyInjection\Compiler;
 
 use FOS\OAuthServerBundle\DependencyInjection\Compiler\TokenStorageCompilerPass;
@@ -8,11 +19,11 @@ use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
- * Class TokenStorageCompilerPassTest
- * @package FOS\OAuthServerBundle\Tests\DependencyInjection\Compiler
+ * Class TokenStorageCompilerPassTest.
+ *
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class TokenStorageCompilerPassTest extends \PHPUnit_Framework_TestCase
+class TokenStorageCompilerPassTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var TokenStorageCompilerPass
