@@ -1,16 +1,25 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the FOSOAuthServerBundle package.
+ *
+ * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace FOS\OAuthServerBundle\Tests;
 
 use FOS\OAuthServerBundle\DependencyInjection\Compiler;
-use FOS\OAuthServerBundle\DependencyInjection\FOSOAuthServerExtension;
 use FOS\OAuthServerBundle\DependencyInjection\Security\Factory\OAuthFactory;
 use FOS\OAuthServerBundle\FOSOAuthServerBundle;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\HttpKernel\Kernel;
 
-class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
+class FOSOAuthServerBundleTest extends \PHPUnit\Framework\TestCase
 {
     protected function setUp()
     {
@@ -20,28 +29,6 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
     public function testConstruction()
     {
         $bundle = new FOSOAuthServerBundle();
-
-        $objectReflection = new \ReflectionObject($bundle);
-
-        $propertyReflection = $objectReflection->getProperty('extension');
-        $propertyReflection->setAccessible(true);
-
-        $this->assertInstanceOf(FOSOAuthServerExtension::class, $propertyReflection->getValue($bundle));
-
-        $propertyReflection = $objectReflection->getProperty('kernelVersion');
-        $propertyReflection->setAccessible(true);
-
-        $this->assertEquals(Kernel::VERSION, $propertyReflection->getValue($bundle));
-    }
-
-    public function testBuildForSymfonyHigherThan20()
-    {
-        $bundle = new FOSOAuthServerBundle();
-        $objectReflection = new \ReflectionObject($bundle);
-
-        $propertyReflection = $objectReflection->getProperty('kernelVersion');
-        $propertyReflection->setAccessible(true);
-        $propertyReflection->setValue($bundle, '2.1.0');
 
         /** @var ContainerBuilder|\PHPUnit_Framework_MockObject_MockObject $containerBuilder */
         $containerBuilder = $this->getMockBuilder(ContainerBuilder::class)
@@ -71,42 +58,6 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
             ->method('addSecurityListenerFactory')
             ->with(new OAuthFactory())
             ->willReturn(null)
-        ;
-
-        $containerBuilder
-            ->expects($this->at(1))
-            ->method('addCompilerPass')
-            ->withConsecutive(
-                new Compiler\GrantExtensionsCompilerPass(),
-                new Compiler\TokenStorageCompilerPass(),
-                new Compiler\RequestStackCompilerPass()
-            )
-            ->willReturnOnConsecutiveCalls(
-                $containerBuilder,
-                $containerBuilder,
-                $containerBuilder
-            )
-        ;
-
-        $this->assertNull($bundle->build($containerBuilder));
-    }
-
-    public function testBuildForSymfony20()
-    {
-        $bundle = new FOSOAuthServerBundle();
-        $objectReflection = new \ReflectionObject($bundle);
-
-        $propertyReflection = $objectReflection->getProperty('kernelVersion');
-        $propertyReflection->setAccessible(true);
-        $propertyReflection->setValue($bundle, '2.0.0');
-
-        /** @var ContainerBuilder|\PHPUnit_Framework_MockObject_MockObject $containerBuilder */
-        $containerBuilder = $this->getMockBuilder(ContainerBuilder::class)
-            ->disableOriginalConstructor()
-            ->setMethods([
-                'addCompilerPass',
-            ])
-            ->getMock()
         ;
 
         $containerBuilder
