@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,16 +13,16 @@
 
 namespace FOS\OAuthServerBundle\Tests\Security\Authentication\Provider;
 
+use FOS\OAuthServerBundle\Model\AccessToken;
 use FOS\OAuthServerBundle\Security\Authentication\Provider\OAuthProvider;
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
-use FOS\OAuthServerBundle\Model\AccessToken;
 use OAuth2\OAuth2;
 use Symfony\Component\Security\Core\Role\Role;
 use Symfony\Component\Security\Core\User\UserCheckerInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 
-class OAuthProviderTest extends \PHPUnit_Framework_TestCase
+class OAuthProviderTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var \PHPUnit_Framework_MockObject_MockObject|UserInterface
@@ -59,13 +61,13 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         ;
         $this->serverService = $this->getMockBuilder(OAuth2::class)
             ->disableOriginalConstructor()
+            ->setMethods(['verifyAccessToken'])
             ->getMock()
         ;
         $this->userChecker = $this->getMockBuilder(UserCheckerInterface::class)
             ->disableOriginalConstructor()
             ->getMock()
         ;
-
         $this->provider = new OAuthProvider($this->userProvider, $this->serverService, $this->userChecker);
     }
 
@@ -76,7 +78,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
 
         $this->user->expects($this->once())
             ->method('getRoles')
-            ->will($this->returnValue(array('ROLE_USER')));
+            ->will($this->returnValue(['ROLE_USER']))
+        ;
 
         $accessToken = new AccessToken();
         $accessToken->setUser($this->user);
@@ -84,17 +87,18 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
         $this->assertSame($this->user, $result->getUser());
-        $this->assertEquals($token->getToken(), $result->getToken());
+        $this->assertSame($token->getToken(), $result->getToken());
         $this->assertTrue($result->isAuthenticated());
         $this->assertCount(1, $result->getRoles());
 
         $roles = $result->getRoles();
-        $this->assertEquals('ROLE_USER', $roles[0]->getRole());
+        $this->assertSame('ROLE_USER', $roles[0]->getRole());
     }
 
     public function testAuthenticateReturnsTokenIfValidEvenIfNullData()
@@ -107,7 +111,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -127,7 +132,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -137,9 +143,9 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $roles = $result->getRoles();
         $this->assertCount(2, $roles);
         $this->assertInstanceOf(Role::class, $roles[0]);
-        $this->assertEquals('ROLE_FOO', $roles[0]->getRole());
+        $this->assertSame('ROLE_FOO', $roles[0]->getRole());
         $this->assertInstanceOf(Role::class, $roles[1]);
-        $this->assertEquals('ROLE_BAR', $roles[1]->getRole());
+        $this->assertSame('ROLE_BAR', $roles[1]->getRole());
     }
 
     public function testAuthenticateWithNullScope()
@@ -153,7 +159,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -175,7 +182,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
