@@ -1,16 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the FOSOAuthServerBundle package.
+ *
+ * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace FOS\OAuthServerBundle\Tests\Util;
 
 use FOS\OAuthServerBundle\Util\Random;
 use phpmock\phpunit\PHPMock;
 
 /**
- * Class RandomTest
+ * Class RandomTest.
  *
  * @author Nikola Petkanski <nikola@petkanski.com
  */
-class RandomTest extends \PHPUnit_Framework_TestCase
+class RandomTest extends \PHPUnit\Framework\TestCase
 {
     use PHPMock;
 
