@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -17,9 +19,9 @@ use FOS\OAuthServerBundle\Model\ClientInterface;
 use FOS\OAuthServerBundle\Model\ClientManagerInterface;
 use OAuth2\OAuth2;
 use OAuth2\OAuth2ServerException;
+use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\Request;
@@ -40,14 +42,13 @@ use Symfony\Component\Security\Core\User\UserInterface;
 class AuthorizeController implements ContainerAwareInterface
 {
     /**
-     * @var ClientInterface
-     */
-    private $client;
-
-    /**
      * @var ContainerInterface
      */
     protected $container;
+    /**
+     * @var ClientInterface
+     */
+    private $client;
 
     /**
      * @var SessionInterface
@@ -105,18 +106,9 @@ class AuthorizeController implements ContainerAwareInterface
     private $eventDispatcher;
 
     /**
-     * Sets the container.
-     *
-     * @param ContainerInterface|null $container A ContainerInterface instance or null
-     */
-    public function setContainer(ContainerInterface $container = null)
-    {
-        $this->container = $container;
-    }
-
-    /**
      * This controller had been made as a service due to support symfony 4 where all* services are private by default.
      * Thus, there is considered a bad practice to fetch services directly from container.
+     *
      * @todo This controller could be refactored to do not rely on so many dependencies
      *
      * @param RequestStack           $requestStack
@@ -158,6 +150,16 @@ class AuthorizeController implements ContainerAwareInterface
     }
 
     /**
+     * Sets the container.
+     *
+     * @param ContainerInterface|null $container A ContainerInterface instance or null
+     */
+    public function setContainer(ContainerInterface $container = null)
+    {
+        $this->container = $container;
+    }
+
+    /**
      * Authorize.
      */
     public function authorizeAction(Request $request)
@@ -176,6 +178,7 @@ class AuthorizeController implements ContainerAwareInterface
         $form = $this->authorizeForm;
         $formHandler = $this->authorizeFormHandler;
 
+        /** @var OAuthEvent $event */
         $event = $this->eventDispatcher->dispatch(
             OAuthEvent::PRE_AUTHORIZATION_PROCESS,
             new OAuthEvent($user, $this->getClient())
@@ -193,10 +196,10 @@ class AuthorizeController implements ContainerAwareInterface
 
         return $this->templating->renderResponse(
             'FOSOAuthServerBundle:Authorize:authorize.html.'.$this->templateEngineType,
-            array(
-                'form'   => $form->createView(),
+            [
+                'form' => $form->createView(),
                 'client' => $this->getClient(),
-            )
+            ]
         );
     }
 
@@ -226,7 +229,8 @@ class AuthorizeController implements ContainerAwareInterface
 
         try {
             return $this->oAuth2Server
-                ->finishClientAuthorization($formHandler->isAccepted(), $user, $request, $formHandler->getScope());
+                ->finishClientAuthorization($formHandler->isAccepted(), $user, $request, $formHandler->getScope())
+            ;
         } catch (OAuth2ServerException $e) {
             return $e->getHttpResponse();
         }
