@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,27 +13,19 @@
 
 namespace FOS\OAuthServerBundle;
 
-use FOS\OAuthServerBundle\DependencyInjection\Compiler\TokenStorageCompilerPass;
+use FOS\OAuthServerBundle\DependencyInjection\Compiler\GrantExtensionsCompilerPass;
 use FOS\OAuthServerBundle\DependencyInjection\Compiler\RequestStackCompilerPass;
+use FOS\OAuthServerBundle\DependencyInjection\Compiler\TokenStorageCompilerPass;
 use FOS\OAuthServerBundle\DependencyInjection\FOSOAuthServerExtension;
 use FOS\OAuthServerBundle\DependencyInjection\Security\Factory\OAuthFactory;
-use FOS\OAuthServerBundle\DependencyInjection\Compiler\GrantExtensionsCompilerPass;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
-use Symfony\Component\HttpKernel\Kernel;
 
 class FOSOAuthServerBundle extends Bundle
 {
-    /**
-     * @example '2.1.0'
-     * @var string
-     */
-    private $kernelVersion;
-
     public function __construct()
     {
-        $this->kernelVersion = Kernel::VERSION;
         $this->extension = new FOSOAuthServerExtension();
     }
 
@@ -39,11 +33,9 @@ class FOSOAuthServerBundle extends Bundle
     {
         parent::build($container);
 
-        if (version_compare($this->kernelVersion, '2.1', '>=')) {
-            /** @var SecurityExtension $extension */
-            $extension = $container->getExtension('security');
-            $extension->addSecurityListenerFactory(new OAuthFactory());
-        }
+        /** @var SecurityExtension $extension */
+        $extension = $container->getExtension('security');
+        $extension->addSecurityListenerFactory(new OAuthFactory());
 
         $container->addCompilerPass(new GrantExtensionsCompilerPass());
         $container->addCompilerPass(new TokenStorageCompilerPass());
